+++ src/app/app.module.ts
@@ -2,21 +2,14 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppRoutingModule } from './app-routing.module';
+
 import { AppComponent } from './app.component';
-import { TitleComponent } from './title/title.component';
-import { PathComponent } from './path/path.component';
-import { DateComponent } from './date/date.component';
-import { MessageComponent } from './message/message.component';
 import { StartButtonsComponent } from './start-buttons/start-buttons.component';
 import { HeaderComponent } from './header/header.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    TitleComponent,
-    PathComponent,
-    DateComponent,
-    MessageComponent,
     StartButtonsComponent,
     HeaderComponent
   ],
