--- src/app/date/date.component.css
@@ -1,3 +1,0 @@
-p {
-  font-size: 25px;
-}
