--- src/app/date/date.component.html
@@ -1,1 +1,0 @@
-<p>{{strDatetime}}</p>
