+++ src/app/header/header.component.css
@@ -5,27 +5,42 @@ header {
   grid-template-rows: repeat(4, 1fr);      /* 1 - 5 */
 }
 
-app-title {
+#app-title {
   text-align: center;
 /*  background-color: red;   */
   grid-column: 1 / 4;
   grid-row: 1 / 3;
 }
-app-path {
+
+#app-path {
 /*  background-color: green;   */
   grid-column: 1 / 3;
   grid-row: 3 / 4;
 }
 
-app-date {
+#app-date {
   text-align: center;
 /*  background-color: lime;   */
   grid-column: 3 / 4;
   grid-row: 3 / 4;
 }
 
-app-message {
+#app-message {
 /*  background-color: yellow;   */
   grid-column: 1 / 4;
   grid-row: 4 / 5;
+}
+
+p {
+  font-size: 25px;
+}
+
+#app-title > p {
+    font-size: 60px;
+    margin-top: 30px;
+    margin-bottom: 30px;
+}
+
+#app-path > p , #app-message > p {
+  padding-left: 42px;
 }
