+++ src/app/header/header.component.ts
@@ -11,6 +11,8 @@ export class HeaderComponent implements OnInit {
   // Auslesen des Input
   @Input('objHeaderData') objHeaderData : HeaderData;
 
+  // Membervariable für Datum
+  strDatetime : string;
 
   constructor() {
 
@@ -18,7 +20,45 @@ export class HeaderComponent implements OnInit {
 
 
   ngOnInit() {
+    this.setStrDateTime();
+  }
+
+  setStrDateTime() {
+
+    let letCurrentDate = new Date();
+    let nDay: number = letCurrentDate.getDay().valueOf();
+
+    let strDayname = "";
+    switch(nDay) {
+      case 0: {
+        strDayname = "Sonntag";       break;
+      }
+      case 1: {
+        strDayname = "Montag";        break;
+      }
+      case 2: {
+        strDayname = "Dienstag";      break;
+      }
+      case 3: {
+        strDayname = "Mittwoch";      break;
+      }
+      case 4: {
+        strDayname = "Donnerstag";    break;
+      }
+      case 5: {
+        strDayname = "Freitag";       break;
+      }
+      case 6: {
+        strDayname = "Samstag";       break;
+      }
+      default: {
+        strDayname = "Fehler beim Auflösen des Wochentags";   break;
+      }
+    }
+
+    this.strDatetime = strDayname + ', der ' + letCurrentDate.toLocaleDateString('de-DE');
 
   }
+
 
 }
