--- src/app/message/message.component.css
@@ -1,4 +1,0 @@
-p {
-  font-size: 25px;
-  padding-left: 40px;
-}
