--- src/app/message/message.component.html
@@ -1,1 +1,0 @@
-<p>{{strMessage}}</p>
