--- src/app/message/message.component.ts
@@ -1,18 +1,0 @@
-import { Component, OnInit, Input } from '@angular/core';
-
-@Component({
-  selector: 'app-message',
-  templateUrl: './message.component.html',
-  styleUrls: ['./message.component.css']
-})
-
-export class MessageComponent implements OnInit {
-
-  @Input('message') strMessage : string;
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
