--- src/app/path/path.component.css
@@ -1,4 +1,0 @@
-p {
-  font-size: 25px;
-  padding-left: 40px;
-}
