--- src/app/path/path.component.html
@@ -1,1 +1,0 @@
-<p>{{strPath}}</p>
