--- src/app/path/path.component.ts
@@ -1,18 +1,0 @@
-import { Component, OnInit, Input } from '@angular/core';
-
-@Component({
-  selector: 'app-path',
-  templateUrl: './path.component.html',
-  styleUrls: ['./path.component.css']
-})
-
-export class PathComponent implements OnInit {
-
-  @Input('path') strPath : string;
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
