--- src/app/title/title.component.css
@@ -1,5 +1,0 @@
-p {
-  font-size: 60px;
-  margin-top: 30px;
-  margin-bottom: 30px;
-}
