--- src/app/title/title.component.html
@@ -1,1 +1,0 @@
-<p>{{strTitle}}</p>
