--- src/app/title/title.component.ts
@@ -1,18 +1,0 @@
-import { Component, OnInit, Input } from '@angular/core';
-
-@Component({
-  selector: 'app-title',
-  templateUrl: './title.component.html',
-  styleUrls: ['./title.component.css']
-})
-
-export class TitleComponent implements OnInit {
-
-  @Input('title') strTitle : string;
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
